Flyttet kode for å hente aktive oppgaver/vurderinger til egen service
