+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -6,6 +6,7 @@ import io.ktor.server.config.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
 import no.nav.syfo.application.cache.RedisStore
+import no.nav.syfo.personstatus.PersonoversiktOppgaverService
 import no.nav.syfo.personstatus.PersonoversiktStatusService
 import no.nav.syfo.personstatus.api.v2.apiModule
 import no.nav.syfo.personstatus.api.v2.auth.getWellKnown
@@ -95,11 +96,13 @@ fun main() {
             personoversiktStatusService = PersonoversiktStatusService(
                 database = database,
                 pdlClient = pdlClient,
-                arbeidsuforhetvurderingClient = arbeidsuforhetvurderingClient,
+                personOversiktOppgaverService = PersonoversiktOppgaverService(
+                    oppfolgingsoppgaveClient = oppfolgingsoppgaveClient,
+                    aktivitetskravClient = aktivitetskravClient,
+                    manglendeMedvirkningClient = manglendeMedvirkningClient,
+                    arbeidsuforhetvurderingClient = arbeidsuforhetvurderingClient,
+                ),
                 personoversiktStatusRepository = personoversiktStatusRepository,
-                oppfolgingsoppgaveClient = oppfolgingsoppgaveClient,
-                aktivitetskravClient = aktivitetskravClient,
-                manglendeMedvirkningClient = manglendeMedvirkningClient,
             )
             personBehandlendeEnhetService = PersonBehandlendeEnhetService(
                 database = database,
