+++ src/main/kotlin/no/nav/syfo/personstatus/api/v2/endpoints/PersonOversiktStatusApiV2.kt
@@ -63,7 +63,7 @@ fun Route.registerPersonoversiktApiV2(
                                 callId = callId,
                                 personOversiktStatusList = personer,
                             )
-                            val personOversiktStatusDTO = personoversiktStatusService.getAktiveVurderinger(
+                            val personOversiktStatusDTO = personoversiktStatusService.getAktiveOppgaver(
                                 callId = callId,
                                 token = token,
                                 personStatusOversikt = personerWithName,
