+++ src/test/kotlin/no/nav/syfo/aktivitetskravvurdering/kafka/AktivitetskravVurderingConsumerSpek.kt
@@ -5,10 +5,6 @@ import io.mockk.*
 import no.nav.syfo.aktivitetskravvurdering.domain.AktivitetskravStatus
 import no.nav.syfo.oppfolgingstilfelle.kafka.toPersonOversiktStatus
 import no.nav.syfo.personstatus.PersonoversiktStatusService
-import no.nav.syfo.personstatus.application.aktivitetskrav.IAktivitetskravClient
-import no.nav.syfo.personstatus.application.arbeidsuforhet.IArbeidsuforhetvurderingClient
-import no.nav.syfo.personstatus.application.manglendemedvirkning.IManglendeMedvirkningClient
-import no.nav.syfo.personstatus.application.oppfolgingsoppgave.IOppfolgingsoppgaveClient
 import no.nav.syfo.personstatus.db.*
 import no.nav.syfo.personstatus.domain.PersonOversiktStatus
 import no.nav.syfo.personstatus.infrastructure.database.repository.PersonOversiktStatusRepository
@@ -34,10 +30,7 @@ class AktivitetskravVurderingConsumerSpek : Spek({
         val personoversiktStatusService = PersonoversiktStatusService(
             database = database,
             pdlClient = externalMockEnvironment.pdlClient,
-            arbeidsuforhetvurderingClient = mockk<IArbeidsuforhetvurderingClient>(),
-            manglendeMedvirkningClient = mockk<IManglendeMedvirkningClient>(),
-            oppfolgingsoppgaveClient = mockk<IOppfolgingsoppgaveClient>(),
-            aktivitetskravClient = mockk<IAktivitetskravClient>(),
+            personOversiktOppgaverService = mockk(),
             personoversiktStatusRepository = personOppgaveRepository,
         )
         val aktivitetskravVurderingConsumer =
