+++ src/test/kotlin/no/nav/syfo/personstatus/infrastructure/kafka/ArbeidsuforhetvurderingConsumerSpek.kt
@@ -6,9 +6,7 @@ import io.mockk.every
 import io.mockk.mockk
 import no.nav.syfo.personstatus.domain.PersonIdent
 import no.nav.syfo.personstatus.PersonoversiktStatusService
-import no.nav.syfo.personstatus.application.aktivitetskrav.IAktivitetskravClient
 import no.nav.syfo.personstatus.application.arbeidsuforhet.IArbeidsuforhetvurderingClient
-import no.nav.syfo.personstatus.application.manglendemedvirkning.IManglendeMedvirkningClient
 import no.nav.syfo.personstatus.application.oppfolgingsoppgave.IOppfolgingsoppgaveClient
 import no.nav.syfo.personstatus.db.createPersonOversiktStatus
 import no.nav.syfo.personstatus.db.getPersonOversiktStatusList
@@ -37,11 +35,8 @@ class ArbeidsuforhetvurderingConsumerSpek : Spek({
     val personoversiktStatusService = PersonoversiktStatusService(
         database = database,
         pdlClient = externalMockEnvironment.pdlClient,
-        arbeidsuforhetvurderingClient = arbeidsuforhervurderingClient,
-        manglendeMedvirkningClient = mockk<IManglendeMedvirkningClient>(),
+        personOversiktOppgaverService = mockk(),
         personoversiktStatusRepository = personOppgaveRepository,
-        oppfolgingsoppgaveClient = oppfolgingsoppgaveClient,
-        aktivitetskravClient = mockk<IAktivitetskravClient>(),
     )
 
     val arbeidsuforhetvurderingConsumer = ArbeidsuforhetvurderingConsumer(
